requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
requests
geopy
coverage
pandas
openpyxl